Reworked epochs table
Two unused dates were removed from the epoch_row, and a completed flag as well as the list of eligible oracles for the epoch have been added.

At the beginning of each epoch, the oracles from the oracle table will be embedded in the epoch as the oracles committing. Administrative actions can be taken to add/remove oracles from specific epochs, in the event an oracle fails to commit or reveal, preventing the epoch from completing.

A newly added oracle will be eligible to commit once the epoch advances.